Fine-tuning register/login screen
