+++ TuffyFood/Controllers/welcomeViewController.swift
@@ -1,0 +1,30 @@
+//
+//  welcomeViewController.swift
+//  TuffyFood
+//
+//  Created by csuftitan on 3/6/22.
+//
+
+import UIKit
+
+class welcomeViewController: UIViewController {
+
+    @IBAction func registerPressed(_ sender: UIButton) {
+    }
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
