+++ TuffyFood/Helpers/Constants.swift
@@ -10,6 +10,8 @@ import Foundation
 struct Constants {
     struct Storyboard {
        static let homeScreenViewController = "HomeVC"
+       static let logInViewController = "LogInVC"
+       static let registerViewController = "registerVC"
        static let authViewController = "AuthVC"
     }
 }
